style : Typography props underlineColor 추가 #4 chore : detailStore Menu 컴포넌트 분리
