--- apps/customer/src/apps/detail-store/components/common/index.ts
@@ -1,1 +1,0 @@
-export { FlexBox } from './FlexBox';
