+++ apps/customer/src/apps/detail-store/components/index.ts
@@ -1,2 +1,3 @@
 export { StoreImage } from './StoreImage';
 export { Info } from './Info';
+export { Menu } from './menu/Menu';
