+++ apps/customer/src/apps/detail-store/components/menu/components/Counter.tsx
@@ -39,7 +39,6 @@ const Wrapper = styled.div`
 
 const Button = styled.button`
   border: none;
-  cursor: pointer;
   background: transparent;
   outline: none;
   /* :hover {
