+++ apps/customer/src/apps/shopping-basket/ShoppingBasket.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import styled from 'styled-components';
+
+export const ShoppingBasket = () => {
+  return (
+    <Container>ShoppingBasket</Container>
+  )
+}
+
+const Container = styled.div`
+  display: flex;
+`;
