+++ apps/customer/src/pages/shopping-basket.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { ShoppingBasket } from '../apps/shopping-basket/ShoppingBasket'
+
+export default function ShoppingBasketPage() {
+  return (
+    <ShoppingBasket />
+  )
+}
